Added flux to likelihood; tidied up code, notebook
Now calculate the inferred flux based on the mdot for each model, and
compare that to the observed flux, and incorporate into the likelihood
calculation.

Included the table from the burst paper, so you can read in the
observational parameters for each of the lightcurves (should probably
include the lightcurves as well)

Reconciled the filename specification for the two burst classes

Also wrote a couple of routines to easily visualise the results of the
MCMC runs.