Modified the ObservedBurst class
This version is now more robust, and can handle the bursts from SAX
J1808.4-3658 where we don't measure an uncertainty on the recurrence time
(because it's assumed to be the separation).

The code also now accepts kwargs similar to KeplerBurst, so that missing
attributes can be included (or existing attributes overwritten).

This also now works with observed bursts NOT in the reference sample.