Minor revisions to the code; started writing companion paper
Added functions to calculate the surface gravity given M, R, both for GR
and in the Newtonian case; also a function to calculate the redshift

Moved the plot section in the compare method for the ObservedBurst to the
end, to make it easier to read.

Updated the __init__ method for KeplerBurst, to ensure that the set of GR
parameters are more complete. Not sure this is really required