Added write method for Lightcurve, fper function
Various features added to improve the utility for burst simulations; in
the observe method, now copies the flux_err from the observed burst
(from which we get the time bins) to the simulated observation.

Can now also write the lightcurves with some limited header information;
these can be subsequently read in with the same code, if required.

Also split off the calculation of fper (used in the compare method) into a
separate function, so we can use it in other contexts