Converted concord to package and updated scripts
A major rearrangement to allow concord to be structured as a package, and
installed using pip. This modification means that the $CONCORD_PATH
environemnt variable is no longer required; the code can find the data
files required via other methods

The anisotropy function is now part of the diskmodel subpackage.

Also updated the tutorial notebook to make sure the code still works after
all the changes. The data files required are now part of the repo, in the
example_data subdirectory