Added homogenize_params to help ensure compatible parameter types
homogenize_params checks the size of all the passed parameters, and if any of them are a distribution, will convert the others to distributions of the same size (by calling value_to_dist). So far the alpha and mdot methods have been converted to use this routine, and the others should also be converted.

Also added logging (in utils.py) via the logging package, to make the informational and error messages look nicer