Lots of small changes to improve robustness, consistency
Introduced intvl_to_errors to convert the percentiles output from functions to proper 2-sided errors as used in input.
Corrected the way value_to_dist deals with scalars, as well as how it treats units with or without prefactors.
Changed the warning, error messages to use the logger instead
Updates to several functions to improve consistency of behaviour