Updated homogenize_params to also work with the inclination
And updated the dist function to use the new functionality of homogenize params. This avoids the need to generate the inclination array by each function, and streamlines further those functions in which inclination plays a role.

Next have to propagate that change through to all the functions

Also updated dist to return (when fulldist=False) the central value and errors, rather than limits, using intvl_to_errors