Now using homogenize_params also for the inclination, where possible
Previously we had a separate step to generate the inclination, but now this is largely incorporated into homogenize_params. Extended this approach through to the other routines that allow an inclination to be passed, excluding hfrac, which is a little more complicated

New "helper" function _i for hfrac, which will return a single-index slice of an array or distribution, OR a scalar value. This is to help with the calculation of xbar, X_0 etc., when using (potentially) a mix of arrays and scalars on input

Also a bunch of other smaller fixes; alpha now outputs a dict when fulldist=True, for consistency with the other routines; value_to_dist now passes a distribution directly through, even when verbose=False