Added the option to use cumulative statistics in asym_dist
This option relies on empirical fits to transform the cumulative parameters (assumed to be 50th, 16th and 84th percentile) into the equivalents for the ML distribution, which can then be used to generate a distribution of the correct shape. The statistics parameter is also accepted by value_to_dist.

Also simplified the Inferring burster properties notebook, and bumped the version number to 1.0.0 for the first release